fix: added directory structure

+++ index.php
@@ -138,11 +138,14 @@ if ($parts[1] === 'user') {
             case 'POST':
                 if (isset($_POST['uploadBtn']) && $_POST['uploadBtn'] == 'upload') {
                     if ($_POST['action'] === 'add') {
+                  //ask how to check parameters      
                         FileService::addFile($_FILES, $_POST['id_dir']);
-                    }
+                    }else {echo 'Enter parameter action=add';}
                     if ($_POST['action'] === 'move') {
                        FileService::moveFile($_POST['newName'], $_POST['id_dir'], $_POST['id_file']);
-                    }
+                    }else {echo 'Enter parameter action=move';}
+                }else {
+                    echo 'Enter parameter uploadBtn=upload ';
                 }
                 break;
             default:
@@ -162,20 +165,22 @@ if ($parts[1] === 'user') {
                 http_response_code(404);
         }
     } elseif ($id === 'share') {
-        if ($fileId != '') {
-            if ($userId == '') {
-                FileAccessService::loadAccess($fileId);
-            } else {
-                if ($method === 'DELETE') {
-                    FileAccessService::delete($fileId, $userId);
-                    //ask
-                } elseif ($method === 'PUT') {
-                    FileAccessService::save($fileId, $userId);
-                    //ask
-                } else  echo 'Your method is incorrect for this app!';
-                http_response_code(404);
-            }
-        } else echo 'share must have id';
+        if($method === 'GET'){
+            if ($fileId != '') {
+                if ($userId == '') {
+                    FileAccessService::loadAccess($fileId);
+                } else {
+                    if ($method === 'DELETE') {
+                        FileAccessService::delete($fileId, $userId);
+                        //ask
+                    } elseif ($method === 'PUT') {
+                        FileAccessService::save($fileId, $userId);
+                        //ask
+                    } else  echo 'Your method is incorrect for this app!';
+                    http_response_code(404);
+                }
+            } else echo 'share must have id';
+        }
     } else {
         echo 'id must be numeric!';
         http_response_code(404);
